cakeservice/src/main/java/com/waracle/cakeservice/service/RoleServiceImpl.java
package com.waracle.cakeservice.service;

import com.waracle.cakeservice.model.Role;
import com.waracle.cakeservice.repository.RoleRepo;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;

@Service
@Transactional
@RequiredArgsConstructor
public class RoleServiceImpl implements RoleService {
    private final RoleRepo roleRepo;

    @Override
    public Role saveRole(Role role) {
        return roleRepo.save(role);
    }
}
